+++ sqlmath_wasm.js
@@ -1,7 +1,7 @@
 /*jslint-disable*/
 // Copyright (c) 2021 Kai Zhu
 // SPDX-License-Identifier: MIT
-// 2025-09-24T07:21:07+0000
+// 2025-09-27T23:39:07+0000
 (function () {
 "use strict";